requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
tzlocal==5.2
yarl==1.9.4
